Switch mod order add mods on other layers
Add each of the 4 modifiers to the left hand when it is holding a chord
for a layer switch on the right hand. This allows using shift and other
modifiers with symbols and arrow keys.

Consistently use CGAS order for the modifiers on the left hand, even
when they shift keys and fingers to move around the held key. Switch the
home row modifiers on the base layer to also use CGAS / SAGC order. This
puts the often used ctrl key under the pinky, but it's the closest to
what I'm used to so it might be easier to learn.