Move pinky keys down a row
I don't know why I put them on a stretch key in the first place when the
pinky home key is unused... Move amp down a row, and swap it with dollar
so that `^` and `$` are at opposite ends. Move backslash down in the
delimiter layer. Add a tab key under the pinky on the mouse/arrow layer.