Drop some layers
Drop the abbreviation layer. Save `l` for some future layer for now.

Drop the second left hand shorcuts layer. Don't use 3 key chording to
get into a layer, but instead move off the home row. I found that with
the chord to enter the layer only one of the chord keys needs to be
held down, so the layer can be maintained by the thumb alone. Add a few
extra shift keys in places where I might be able to use them with a
pointer finger that isn't holding the layer shift. Consider adding a
layer lock too...