Reduce use of outer columns
Reduce the amount the pinky needs to reach the outer column and bring
the layout closer to compatibility with a 36 key board.

The left and right layer shortcut layers both have the entire other side
unbound so there is lots of room for even closer layer switched
shortcuts. Add a shortcut for ctrl with semicolon and quote. This shifts
the workspace change shorcuts left by 1 key.

Move the control layer shortcuts and RGB mode key to the center column
instead of the outer column.

Reorganize the number layer. Put slash and asterisk on a tapdance key to
free up the outer column. Move minus to the left of the numpad so it
makes sense when entering a negative number. Move dot to the left as
well.