Add some toggleable modifiers
I don't see an easy way to toggle on a single press, setting
`ONE_SHOT_TAP_TOGGLE` to `1` does not work. Set it as low as possible.

Add the one shot mods under the left hand on the layer shift. Can click
once for one shot, and twice to hold.

Make normal mods capital, so one shot mods can be lowercase on the
screen.