Improve layer switch UX
With a combo into a layer it previously required lifting the thumb
before entering into a new layer. This is different than if the layer
switch was done rolling from thumb to finger, where the thumb can stay
down and enter a different layer. Lifting the thumb while the finger
remains had no effect.

Now lifting the thumb immediately goes back to the default layer.
Lifting the finger moves to the shift layer where another layer can be
chosen.