Keep layer when lifting thumb
Return false always so that holding either key in the combo retains the
layer. Previously with the true return the layer would be left
immediately on lifting the thumb. Track when the thumb was lifted first
and don't set the shift layer when releasing the second key.
When rolling the layer is kept when lifting the thumb, so try to keep
the behavior the same between them.