Add RGB layer key hints
Add a map of colors for each key with a function on each layer above the
base layer.
Add a RGB matrix indicator callback which reads the array and assigns
colors for each key in range. Scale the colors by the brightness. Divide
in 2 again when using the heatmap RGB effect since those colors are more
dim. On other effects the brightness appears higher still for the hint
keys, I am not sure why. Using the halved approach makes a close match,
but leaving the hints brighter is more legible when the RGB effect is
noisy.

Increase the ramp rate for the heatmap effect.
React to key releases instead of presses.
Lower RGB effect maximum brightness (even though I'm not sure I noticed
the difference).

Swap the volume up/down so that down is next to mute.

Swap RGB saturation up/down since there effect is more like the opposite
of the value (brightness) changing.